Add files via upload
Initial Commit. The file called matlabSimulation contains MATLAB code used to test the PSO algorithm with Choset and Kim et al. Artificial Potencial Fields in ideal conditions. The file called webotsSimulation contains the Webots worlds and controllers to test in cases A, B and C the implementation of the PSO algorithm and the APFs in the epuck model.